requirements.txt
Django
djangorestframework
django-cors-headers
django-environ
django-extensions
djangorestframework-simplejwt
django-filter
drf-spectacular
python-decouple
django-debug-toolbar
dj-rest-auth
drf-spectacular
drf-spectacular-sidecar
django_q
setuptools
django-activity-stream
requests

wmi
psutil